Linux crosscompilation: Fix missing file due to backslash
This change does not affect the native windows build in any negative way.
This is because the standard demands the compiler accept forwardslash "/" as folder separator and then convert it to whatever is appropiate for the platform.
This goes back to C and the use of "/" as folder separator in UNIX.